Fix icmp duration (#17920)
At some point in the past we apparently stopped nesting [ICMP
fields](https://www.elastic.co/guide/en/beats/heartbeat/master/exported-fields-icmp.html).
This patch fixes that issue.

(cherry picked from commit 754eac174cdb2a335ce578c39130b163b40b49d1)
